Adds 'Unsafe' item serialization / deserialization for metadata.
