+++ build.gradle
@@ -1,7 +1,7 @@
 plugins {
     id 'java'
-    id 'org.jetbrains.kotlin.jvm' version '1.3.72'
-    id "net.kyori.indra" version "2.0.6"
+    id 'application'
+    id'org.jetbrains.kotlin.jvm' version '1.3.72'
 }
 
 group 'near.vanilla.death.database'
@@ -10,27 +10,17 @@ version '1.0'
 repositories {
     google()
     mavenCentral()
-    maven {
-        url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/'
-
-        // As of Gradle 5.1, you can limit this to only those
-        // dependencies you expect from it
-        content {
-            includeGroup 'org.bukkit'
-            includeGroup 'org.spigotmc'
-        }
-    }
+    mavenLocal()
     maven { url = 'https://oss.sonatype.org/content/repositories/snapshots' }
     maven { url = 'https://oss.sonatype.org/content/repositories/central' }
     maven { url = "https://jitpack.io" }
     maven { url = "https://libraries.minecraft.net" }
-    maven { url = 'https://papermc.io/repo/repository/maven-public/' }
+    maven { url "https://papermc.io/repo/repository/maven-public/" }
 }
 
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    // Pick only one of these and read the comment in the repositories block.
-    compileOnly 'org.spigotmc:spigot-api:1.17-R0.1-SNAPSHOT' // The Spigot API with no shadowing. Requires the OSS repo.
+    implementation "org.jetbrains.kotlin:kotlin-stdlib"
+    implementation "io.papermc.paper:paper-api:1.17.1-R0.1-SNAPSHOT"
     implementation 'net.kyori:adventure-api:4.8.1'
     implementation("net.kyori:adventure-text-minimessage:4.1.0-SNAPSHOT")
 
@@ -53,6 +43,7 @@ compileKotlin {
 compileTestKotlin {
     kotlinOptions.jvmTarget = "1.8"
 }
+
 
 jar {
     duplicatesStrategy = DuplicatesStrategy.EXCLUDE
