+++ settings.gradle
@@ -1,1 +1,2 @@
 rootProject.name = 'deathdatabase'
+
