--- src/main/kotlin/near/vanilla/death/database/ItemMetaSerializer.kt
@@ -1,7 +1,0 @@
-package near.vanilla.death.database
-
-class ItemMetaSerializer {
-
-
-
-}
